minor refactor for screen 2 code
