+++ lib/myHomePage.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'progress.dart';
 import 'taskList.dart';
 import 'button.dart';
+import 'screen2.dart';
 
 class MyHomePage extends StatelessWidget {
 
@@ -31,37 +32,6 @@ MaterialPageRoute(builder: (context) => Screen2()),
             child: Text('Go to Screen 2'),
           ),
           // Screen1(),
-        ],
-      ),
-    );
-  }
-}
-
-class Screen2 extends StatelessWidget {
-  // Function to navigate back to Screen1
-  void navigateBack(BuildContext context) {
-    Navigator.pop(context);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Screen 2'),
-        centerTitle: true,
-      ),
-      body: Column(
-        mainAxisAlignment: MainAxisAlignment.center, // Center the button vertically
-        children: [
-          Center(
-            child: ElevatedButton(
-              onPressed: () {
-                navigateBack(
-                    context); // Use the function to navigate back to Screen1
-              },
-              child: Text('Go Back to Screen 1'),
-            ),
-          ),
         ],
       ),
     );
